project_api/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==2.2.2
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
django-simpleui==2021.8.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.1.0
eventlet==0.31.1
http-client==0.1.22
importlib-metadata==4.6.1
itsdangerous==2.0.1
kombu==5.1.0
lxml==4.6.3
mysqlclient==2.0.3
openpyxl==3.0.7
Pillow==8.3.1
pipreqs==0.4.10
prompt-toolkit==3.0.19
pycparser==2.20
pycryptodomex==3.9.4
pycurl==7.43.0.5
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-alipay-sdk==2.0.0
pytz==2021.1
qcloudsms-py==0.1.4
redis==3.5.3

